src/requirements.txt
argon2-cffi
requests
